hexfloat2: ieee754 floating point numbers in hex format
Add a parser and a formatter for the "hexfloat" format for f32 and f64.
The formatter always uses a precise format tied to the exact bit
representation; the parser is forgiving and tries to accept all of the
inputs that other languages allow, e.g. 0x1.0 == 0x1p0 == 0x.1p1 .